### 2021-03-15 #3113 @whedon
[PRE REVIEW]: Analysis of Prehistoric Iconography with the R package iconr
**Submitting author:** @zoometh (<a href="http://orcid.org/0000-0002-1112-6122">Huet Thomas</a>)
**Repository:** <a href="https://github.com/zoometh/iconr" target ="_blank">https://github.com/zoometh/iconr</a>
**Version:** v0.1.0
