### 2021-03-22 #3113 @zoometh
@arfon Yes, [iconr](https://cran.r-project.org/web/packages/iconr/index.html) package LOC can < 1000, but the theoretical background is innovative and voluntarily parsimonious because the package should be at the root of different implementations. Best,
