### 2021-03-31 #3113 @kthyng
Hi @zoometh! The editorial board has determined that your submission is in scope.

We will have to ask for your patience, though, as we have a backlog of submissions and our relevant editors are already overbooked with submissions. I will add the waitlist label now.
