### 2021-05-03 #3191 @fbiljecki
Kudos to the authors for developing this software and releasing it openly. Some comments:

- There is no LICENSE file, and the license is not explicit in the readme/documentation
- I cannot replicate the example in the paper before Fig 2 including the function `plot_dec_grph()` (I get `object 'imgs' not found`)
- Please add a paragraph about related efforts/work
