### 2021-05-04 #3191 @hugoledoux
@zoometh the first review (from a GIS perspective) is in and it is a good idea to fix what is missing and update the submission now, you do not need to wait for the other reviewer (archeology perspective).

The part about citing and comparing in the paper to other similar software would be especially important to add.
