+++ doc/articles/paper.bib
@@ -39,3 +39,59 @@ url = {https://archiv.ub.uni-heidelberg.de/propylaeumdok/512/1/11_04_alexander_b
   bth=2172,
   url = {https://eprints.ucm.es/11070/1/T32200.pdf},
 }
+
+@book{SaintMartin11,
+  title={S{\'e}miologie du langage visuel},
+  author={Saint-Martin, Fernande},
+  year={2011},
+  publisher={PUQ},
+  bth=3104
+}
+
+@InProceedings{Chenorkian95,
+  author    = {Chenorkian, Robert},
+  title     = {Le vestige arch{\'e}ologique gravure rupestre: {\'e}tude et interpr{\'e}tation},
+  booktitle = {L'homme m{\'e}diterran{\'e}en (m{\'e}langes offerts {\`a} Gabriel Camps)},
+  year      = {1995},
+  pages     = {157--178},
+  bth       = {907},
+  publi     = "TP",
+}
+
+@book{Saussure89,
+  title={Cours de linguistique g{\'e}n{\'e}rale},
+  author={De Saussure, Ferdinand},
+  volume={1},
+  year={1989},
+  publisher={Otto Harrassowitz Verlag},
+  bth=3102,
+  publi = "TP"
+}
+
+@article{Binford62,
+  title={Archaeology as anthropology},
+  author={Binford, Lewis R},
+  journal={American antiquity},
+  pages={217--225},
+  year={1962},
+  publisher={JSTOR},
+  bth=3360
+}
+
+@book{Gell98,
+  title={Art and agency: an anthropological theory},
+  author={Gell, Alfred},
+  year={1998},
+  publisher={Clarendon Press},
+  bth=3097,
+  publi = "TP"
+}
+
+@book{Hodder82,
+  title={Symbolic and structural archaeology},
+  author={Hodder, Ian and others},
+  year={1982},
+  publisher={Cambridge University Press Cambridge},
+  bth=3361
+}
+
