### 2021-05-07 #3191 @zoometh
@hugoledoux:

> The part about citing and comparing in the paper to other similar software would be especially important to add.

We have made the required changes. The R package iconr is the first software that gives a parsimonious manner to model 'graphical units' and 'graphical compositions'. And, as a consequence, that permits to conduct cross-cultural and over the long-term comparisons with a minimal loss of information. Since not such a tool exists before, we insist on the theoretical background and the historiography of Archaeology.

@fbiljecki:

> There is no LICENSE file, and the license is not explicit in the readme/documentation

We have added the GPL-3 license ([https://github.com/zoometh/iconr/blob/master/LICENSE](https://github.com/zoometh/iconr/blob/master/LICENSE))

> I cannot replicate the example in the paper before Fig 2 including the function plot_dec_grph() (I get object 'imgs' not found)

We have added the missing code lines before the `plot_dec_grph()` call:

```
nds.df <- read_nds(site, decor, dataDir)
eds.df <- read_eds(site, decor, dataDir)
imgs <- read.table(paste0(dataDir, "/imgs.tsv"),
                   sep="\t", stringsAsFactors = FALSE)
plot_dec_grph(nds.df, eds.df, imgs,
              site, decor, dataDir)
```

Please let me know if additional changes are needed
