### 2021-05-10 #3191 @fbiljecki
I confirm that I can now replicate the example code. Expecting that the same code will be included in the updated version of the paper, I do not have further comments.
