### 2021-05-14 zoometh/iconr#73 @benmarwick
add community guidelines documents
As part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/3191#issuecomment-841344707), I have noticed that the project is lacking **Community guidelines.** We do not see "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I recommend to add
- [x] a CONTRIBUTING.md document (e.g. https://github.com/tidyverse/tidyverse/blob/master/.github/CONTRIBUTING.md) and
- [x] a CODE_OF_CONDUCT.md document (e.g. https://github.com/tidyverse/tidyverse/blob/master/.github/CODE_OF_CONDUCT.md) and
- [x] Issue templates (e.g. https://github.com/tidyverse/tidyverse/tree/master/.github/ISSUE_TEMPLATE)
