### 2021-05-16 zoometh/iconr#73 @zoometh
@benmarwick I've added the missing guidelines documents:

- [CONTRIBUTING.md](https://github.com/zoometh/iconr/blob/master/.github/CONTRIBUTING.md)
- [CODE_OF_CONDUCT.md](https://github.com/zoometh/iconr/blob/master/.github/CODE_OF_CONDUCT.md)
- [Issue templates](https://github.com/zoometh/iconr/blob/master/.github/ISSUE_TEMPLATE.md)
