### 2021-05-17 #3191 @hugoledoux
thanks @benmarwick

I guess you didn't accept the invite a while ago and it wasn't valid anymore. I re-invited you, but also clicked the checkboxes on your behalf above.

> Really wonderful to see this, and I am keen to employ it in my research.

Very nice to read this, this is the whole point of JOSS!
