### 2021-05-17 #3191 @zoometh
@hugoledoux As far as I understand, the [only issue](https://github.com/zoometh/iconr/issues/73) was about missing docs for contributors. I've added these docs and closed the issue. Is there's something I've forgotten?
