### 2021-05-17 #3191 @hugoledoux
oh my bad, I had not noticed you already answered, sorry.

I have fixed a few things in the paper/bib and made 2 PR:
https://github.com/zoometh/iconr/pull/74
https://github.com/zoometh/iconr/pull/75

and I have 3 things I am not sure about:

  1. the DOI checker complains about one book but not the others. I am not sure if all old books have DOIs, but we like DOIs at JOSS so could you check if all the books you cite have one? If you you add it?
  2. l.40: speaks of "Harris diagram", and a a layman I have no idea what those are. A reference perhaps?
  3. l.50: (birel: touches): here I have no clue what "birel" means. Perhaps it's me? Could you explain here and eventually in the text?
