### 2021-05-17 7d46db8
paper varia
