+++ doc/articles/paper.bib
@@ -92,6 +92,14 @@ url = {https://archiv.ub.uni-heidelberg.de/propylaeumdok/512/1/11_04_alexander_b
   author={Hodder, Ian and others},
   year={1982},
   publisher={Cambridge University Press Cambridge},
+  doi = {10.1017/CBO9780511558252},
   bth=3361
+}
+
+@book{Harris14,
+  title={Principles of archaeological stratigraphy},
+  author={Harris, Edward C},
+  year={2014},
+  publisher={Elsevier}
 }
 
