+++ doc/articles/paper.md
@@ -43,7 +43,7 @@ Even the reevaluation of semiotics paradigms following the scientific trends  --
 
 # Core functionality
 
-The R package `iconr` is designed to offer a greater normalization of quantitative indexes for iconography studies [@Alexander08; @Huet18a]. It is grounded in graph theory and spatial analysis to offer concepts and functions for modeling prehistoric iconographic compositions and preparing them for further analysis (clustering, typology tree, Harris diagram, etc.). The main principle of the `iconr` package is to consider any iconographic composition (here, 'decoration') as a geometric graph of graphical units. Geometric graphs, also known as *planar graphs* or *spatialized graphs*, allow to model the neighborhood of these graphical unit which are the fundamental relationships of visual semiotics [@SaintMartin11]. Graphical units are decorated surfaces (`POLYGONS`) modeled as nodes (`POINTS`) and tagged with semantic content (type, color, orientation, etc.). Separable graphical units showing a main graphical content (e.g., type = anthropomorphic figure) are considered as *main* nodes. Graphical units showing a specification of a *main* node (e.g. a sword handed by this anthropomorphic figure) are considered as *attribute* nodes. Each pair of *main* nodes thought to be contemporary that share a border (*birel*: touches) of their Voronoi cells, are connected by an undirected edge (`LINES`).
+The R package `iconr` is designed to offer a greater normalization of quantitative indexes for iconography studies [@Alexander08; @Huet18a]. It is grounded in graph theory and spatial analysis to offer concepts and functions for modeling prehistoric iconographic compositions and preparing them for further analysis: clustering, typology tree, Harris diagram [i.e. temporal succession of archaeological contexts, Harris -@Harris14], etc. The main principle of the `iconr` package is to consider any iconographic composition (here, 'decoration') as a geometric graph of graphical units. Geometric graphs, also known as *planar graphs* or *spatialized graphs*, allow to model the neighborhood of these graphical unit which are the fundamental relationships of visual semiotics [@SaintMartin11]. Graphical units are decorated surfaces (`POLYGONS`) modeled as nodes (`POINTS`) and tagged with semantic content (type, color, orientation, etc.). Separable graphical units showing a main graphical content (e.g., type = anthropomorphic figure) are considered as *main* nodes. Graphical units showing a specification of a *main* node (e.g. a sword handed by this anthropomorphic figure) are considered as *attribute* nodes. Each pair of *main* nodes thought to be contemporary that share a border (binary topological relationship: *touches*) of their Voronoi cells, are connected by an undirected edge (`LINES`).
   
   
 <center>
