### 2021-05-17 #3191 @zoometh
> I have fixed a few things in the paper/bib and made 2 PR:
> zoometh/iconr#74
> zoometh/iconr#75

I've merged your corrections on `paper.bib`

> 1. the DOI checker complains about one book but not the others. I am not sure if all old books have DOIs, but we like DOIs at JOSS so could you check if all the books you cite have one? If you you add it?

I found 1 supplementary DOI (for Hodder's book). I've added it

> 2. l.40: speaks of "Harris diagram", and a a layman I have no idea what those are. A reference perhaps?

Done. I've added a short description of what is a Harris diagram, and a reference

> 3. l.50: (birel: touches): here I have no clue what "birel" means. Perhaps it's me? Could you explain here and eventually in the text?

Done. I've changed "*birel*: touches" to a more explicit "binary topological relationship: *touches*"
