### 2021-05-17 #3191 @zoometh
I've committed these changes and added also minor changes to avoid closing and opening parenthesis touching each other, e.g.: "..)(..".

 I've run the Whedon preview: it seems fine
