### 2021-05-17 #3191 @hugoledoux
OK, both reviewers recommend acceptance so we're moving towards this.

At this point could you:

Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
