+++ NEWS.md
@@ -1,2 +1,5 @@
+# Version 0.1.1 (17 May 2021)
+First release.
+
 # Version 0.1.0 (12 February 2021)
 First official CRAN release.
