### 2021-05-17 #3191 @zoometh
> Make a tagged release of your software, and list the version tag of the archived version here

Done -> [Warrior stelae](https://github.com/zoometh/iconr/releases/tag/0.1.1)

> Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)...
> Please list the DOI of the archived version here.

Done. The DOI of the archive is [10.5281/zenodo.4767529](https://doi.org/10.5281/zenodo.4767529)
